Changes to make multiple masonry containers work in single page
As pointed out in issue #27, multiple masonry elements in a single page
do not work. Changes have been made in the js file to facilitate the
same. I’ve tested this in one of my projects and is working fine. Hence
I thought, I may contribute to this library to make the same available
in the coming release.